+++ stream-chat-android-docs/src/main/kotlin/io/getstream/chat/docs/kotlin/Android.kt
@@ -623,18 +623,19 @@ class Android {
             val chatDomain = ChatDomain.instance()
         }
 
-        fun customizeRetryPolicy() {
-            val chatDomain = ChatDomain.instance()
+        fun initializeChatDomainWithCustomRetryPolicy() {
+            val chatClient = ChatClient.Builder("apiKey", requireContext()).build()
+            val chatDomain = ChatDomain.Builder(requireContext(), chatClient)
+                .retryPolicy(object : RetryPolicy {
+                    override fun shouldRetry(client: ChatClient, attempt: Int, error: ChatError): Boolean {
+                        return attempt < 3
+                    }
 
-            chatDomain.retryPolicy = object : RetryPolicy {
-                override fun shouldRetry(client: ChatClient, attempt: Int, error: ChatError): Boolean {
-                    return attempt < 3
-                }
-
-                override fun retryTimeout(client: ChatClient, attempt: Int, error: ChatError): Int {
-                    return 1000 * attempt
-                }
-            }
+                    override fun retryTimeout(client: ChatClient, attempt: Int, error: ChatError): Int {
+                        return 1000 * attempt
+                    }
+                })
+                .build()
         }
 
         fun watchChannel() {