+++ CHANGELOG.md
@@ -42,8 +42,10 @@
 
 ### ✅ Added
 - Added new extension function `ChatCliet::requestMembers` to query members without `ChatDomain`.
+- Added new extension function `ChatCliet::searchUsersByName`.
 
 ### ⚠️ Changed
+- 🚨 Breaking change: `RetryPolicy` in `ChatDomain` is now immutable and can only be set with Builder before creating an instance of it.
 
 ### ❌ Removed